refactor: widen ComponentId to 64 bits
Reserve room for runtime-created relation-pair ids to share the component id
space. Mechanical: the id is only ever used as a dense index, never packed.

+++ crates/flux_ecs/src/storage/layout.rs
@@ -206,7 +206,7 @@ mod tests {
         let mut out: Vec<ComponentId> = keys
             .iter()
             .map(|k| {
-                (0..r.len() as u32)
+                (0..r.len() as u64)
                     .map(ComponentId)
                     .find(|id| r.info(*id).key == *k)
                     .expect("registered")
